### 2020-02-07 #2085 @whedon
[PRE REVIEW]: mimosa: A Modern Graphical User Interface for 2-level Mixed Models
**Submitting author:** @johannes-titz (<a href="http://orcid.org/0000-0002-1102-5719">Johannes Titz</a>)
**Repository:** <a href="https://github.com/johannes-titz/mimosa" target ="_blank">https://github.com/johannes-titz/mimosa</a>
**Version:** v0.2.0
