### 2020-02-10 #2085 @danielskatz
👋 @cMadan - just a quick ping at the start of a new week and a new AEiC on duty...
