### 2020-02-18 #2085 @johannes-titz
@cMadan Thank you! I would like to suggest Daniel Lüdecke (github: strengejacke). He is an expert in mixed models and R. Otherwise, adibender and fabian-s might be suitable reviewers as they list mixed models in their domain. aj2duncan is another potential reviewer as he is familiar with shiny.

By the way, the reviewers document has some duplicates. They should be easy to identify and remove, which would make the file a bit cleaner. It might also be worth to sort the file by some columns (by default), e.g. programming language--active reviews--review count. This would make the selection of reviewers a bit easier. (Just a suggestion.)
