### 2020-02-18 #2085 @cMadan
@johannes-titz, thanks for the suggestions--both for reviewers and on our process. We are working to revamp how the reviewer list works to make it a bit more dynamic (e.g., active reviews, total count) and that's in progress. :)
