### 2020-03-09 #2116 @aj2duncan
Hi @johannes-titz,

I've got a query for you if that's ok. In terms of contributions you list a contributer on mimosa.org and a second user has contributed 71 commits to the repo. I'm not sure if this is the same person or different people, but perhaps you need to consider the authorship of the submission?

Thanks,

Andrew
