### 2020-03-09 #2116 @strengejacke
Thanks @johannes-titz,
the paper is well written and the website and readme are well documented. Beside @aj2duncan comment that you should think of adding other contributors either as "contributor" to the description-file, or even as co-author if this is justified, I have no further remarks regarding this review.
However, I have a methodological remark. When including level-2 predictors as fixed effects, you might offer "de-meaning" variables, to avoid that the error terms correlate with covariates (which violates the iid-assumption of mixed models). You find some details here: https://easystats.github.io/parameters/reference/demean.html and in the references literature. I would not make this a requirement for this review, but I encourage you to think about considering "group-meaning" your level-2 predictors in your package.
