### 2020-03-12 85f3348
add Maria Reichert as contributor, normalize
