### 2020-03-16 8a0b2ce
fix problem with "estimate model button"
When loading a new file, the model button would disappear since the inputs are dynamically created. Fixed this by using the reactive mode checkbox as a conditional (isolated).
