### 2020-03-16 9d38972
add busy spinner and notification
For long lme4 computations, a notification is not displayed. Further, a spinner is displayed anytime the server is busy.
