### 2020-03-17 58aa5df
filter independent variables
Nominal variables (factors) with more than 10 levels are excluded from the analysis.
