### 2020-03-17 169983b
update local tests
sleepstudy and Exam are already included in the normal tests, so they can be removed. Furthermore, the sorting alogrithm changed, which had an effect on the atemm data set. Another test has been added for atemm to insure that sorting is consistent.
