+++ NAMESPACE
@@ -9,8 +9,10 @@ importFrom(Hmisc,spss.get)
 importFrom(lme4,lmer)
 importFrom(plyr,ldply)
 importFrom(readr,guess_encoding)
+importFrom(shinyBS,bsTooltip)
 importFrom(shinyalert,shinyalert)
 importFrom(shinyalert,useShinyalert)
+importFrom(shinybusy,add_busy_spinner)
 importFrom(shinydashboard,box)
 importFrom(shinydashboard,dashboardBody)
 importFrom(shinydashboard,dashboardHeader)
