+++ DESCRIPTION
@@ -25,6 +25,8 @@ Imports:
     lme4,
     shiny,
     shinyalert,
+    shinyBS,
+    shinybusy,
     shinydashboard,
     shinyjs,
     sjPlot,
@@ -38,4 +40,4 @@ Suggests:
     shinytest
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.0.2
+RoxygenNote: 7.1.0
