### 2020-03-17 b122c2b
use a simple tooltip
shinyBS does not really work, so I simply use tags$div with a title
