### 2020-03-18 #2116 @aj2duncan
@johannes-titz I really like those changes that you've made. The spinner that you've added is just exactly the thing that I was thinking about, rather than something more accurate using the convergence of lme4. I also like the reactive vs non-reactive mode. Very simple to use. Changing between the two settings did also show/hide the contributions, which you may want to tweak.

Otherwise, really like the changes. @cMadan from my point of view - very happy for this review to be concluded.
