### 2020-03-20 #2116 @aj2duncan
@johannes-titz I was using your example app at the joss-review link above on Chrome but now I can't reproduce it either - so nevermind not sure what was happening.
