### 2020-03-27 #2116 @johannes-titz
@aj2duncan Ok. I guess it can happen if there are many variables and the footer is moved down because of this, but I am not really sure.

@cMadan Thank you! If I can help somehow, let me know. The master is now up to date. If you think copy-editing for the paper is necessary (as you can see I am not a native speaker), I will look for someone.
