+++ paper.md
@@ -17,17 +17,17 @@ bibliography: library.bib
 ---
 
 # Summary
-The mixed models special agent (``mimosa``) is a ``shiny`` [@chang2019] app for 2-level mixed models. Mixed models are rapidly becoming the gold standard of statistical analysis techniques in the behavioral sciences. At the same time there exist few user-friendly software to conduct mixed model analyses. The most common tools often lack a graphical user interface, are proprietary, and involve a tedious process of getting data in and publication-ready tables out. The shiny app ``mimosa`` is supposed to offer an alternative that is free, open source, intuitive, and runs in a browser, making it easily accessible (see https://www.mimosa.icu).
+The mixed models special agent (``mimosa``) is a ``shiny`` [@chang2019] app for 2-level mixed models. Mixed models are rapidly becoming the gold standard of statistical analysis techniques in the behavioral sciences. Yet there is only a small number of user-friendly programs for conducting mixed model analyses. The most common tools often lack a graphical user interface, are proprietary, and involve a tedious process of getting data in and publication-ready tables out. An exception is the shiny app ``mimosa`` which offers an alternative that is free, open source, intuitive, and runs in a browser, making it easily accessible (see https://www.mimosa.icu).
 
-The software is targeted at behavioral scientists who frequently use 2-level mixed models and want a tailored solution for this specific use case. For instance, researchers studying groups (e.g. students clustered in schools, individuals clustered in work groups) and researchers employing within-subjects designs almost exclusively analyze their data with 2-level mixed models. In contrast to other software, ``mimosa`` was designed for this use case. It helps the analyst by automatically detecting potential grouping variables and categorizing variables in level 1 and level 2. Furthermore, ``mimosa`` is researcher-oriented because it produces a single summary table via ``sjPlot`` [@ludecke2018] that can be published in a scientific journal without any modifications.
+The software is targeted at behavioral scientists who frequently use 2-level mixed models and want a solution that is tailored for this particular use case. For instance, researchers studying groups (e.g. students clustered in schools, individuals clustered in work groups) and researchers employing within-subjects designs almost exclusively analyze their data with 2-level mixed models. Unlike other software, ``mimosa`` was designed for this use case. It helps the analyst by automatically detecting potential grouping variables and categorizing these variables in level 1 and level 2. Furthermore, ``mimosa`` is researcher-oriented because it produces a single summary table via ``sjPlot`` [@ludecke2018] that can be published in a scientific journal without any modifications.
 
-To my knowledge, there are only few free, open source, software packages for mixed models that have a graphical user interface. For a slightly outdated comparison of different mixed model software (including proprietary and command-line software) see @west2012. ``LMMgui`` focuses on within-subjects-designs and as such can be seen as a direct competitor of ``mimosa``. ``LMMgui`` offers diagnostic plots and can compare two models, which is missing in ``mimosa``. Disadvantages of ``LMMgui`` are that it is only available for Windows, while ``mimosa`` runs on any platform because it is a browser application. Furthermore, ``mimosa`` has a cleaner interface, is easier to use and produces a publication-ready output.
[…]
