### 2020-04-06 4f1bf54
change intro of readme so that the wording is the same as in the paper
