### 2020-04-06 #2116 @johannes-titz
@cMadan Just to keep you up to date: I sent the paper to a copy-editor and she found a couple of small grammar mistakes and made some minor improvements in word choice. These changes are already included in the master (see: https://github.com/johannes-titz/mimosa/commit/a80736f5442ff506cb489d937b542b1fb91f3ad0?short_path=fb6416b). I believe this is now good to go.
