+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: mimosa
 Title: Graphical User Interface for 2-Level Mixed Models as a
     Shiny-App
-Version: 0.2.0.9000
+Version: 0.3.0
 Authors@R: 
     c(person(given = "Johannes",
              family = "Titz",
