### 2020-04-29 #2116 @johannes-titz
@cMadan: Thank you! Here comes the new tagged version and the zenodo archive:
v.0.3.0: https://github.com/johannes-titz/mimosa/tree/v0.3.0
Zenodo doi: 10.5281/zenodo.3775143 (changed title and author list, added ORCID)
